requirements/production.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & NLP
openai==1.3.5
anthropic==0.7.4
groq==0.3.0
langchain==0.0.350
langdetect==1.0.9
qdrant-client==1.7.0

# Communication
twilio==8.10.1
websockets==12.0
python-socketio==5.10.0
aioredis==2.0.1

# WhatsApp
requests==2.31.0
pydantic==2.5.0

# Payments
stripe==7.8.0

# Utils
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pillow==10.1.0
httpx==0.25.2

# Monitoring
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0